+++ components/display/table/TableData.mjs
@@ -9,6 +9,7 @@ export default class TableData extends TableCell {
      * @param {object} props Component properties
      * @param {*} props.sortValue
      * @param {string} props.sortKey
+     * @param {null|Function} props.renderer
      * @param {string[]|HTMLElement[]|AbstractElement[]} children Component children
      * @param {string} theme Component color theme
      * @param {string} id Component ID
@@ -17,14 +18,7 @@ export default class TableData extends TableCell {
      */
     constructor(props = {}, children = [], theme = null, id = null, classes = [], attributes = {}) {
         super(props, children, theme, id, classes, attributes);
-    }
 
-    render() {
-        let el = document.createElement("td");
-
-        this.attachChildren(el);
-        this.attachListeners(el);
-
-        return el;
+        this._tagName = "td";
     }
 }